+++ src/main/java/org/adria/virementMultiple/service/IBanqueService.java
@@ -13,7 +13,7 @@ public interface IBanqueService {
 	List<Beneficiary> getBeneficiaries(Subscriber subscriber);
 	Beneficiary getBeneficiaryById(Long id);
 	List<Account> getAccountBySubscriber(Subscriber subscriber);
-	List<Account> getAccountBySubscriberId(int id);
+	List<Account> getAccountBySubscriberId(Long id);
 	Boolean Save(List<MultipleTransfersBeneficiary> mtb);
 	Boolean Confirme(String username,String password);
